Additional features
I created the scripts to log collisions with the walls and enemies, and I created the script for the collision counter at the top of the window where the collision count will be kept as the player makes his way through the game. I also populated the scene with more walls and many more enemies.